RFPlugin: examples + class separation
