+++ example/RFModulePlugin/moduleLoader/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(SRC main.cpp)
+add_executable(moduleLoader ${SRC})
+target_link_libraries(moduleLoader YARP::YARP_OS
+                                   YARP::YARP_init)
