+++ example/RFModulePlugin/testmoduleBar/CMakeLists.txt
@@ -1,0 +1,25 @@
+################################################################################
+#                                                                              #
+# Copyright (C) 2018 Fondazione Istituto Italiano di Tecnologia (IIT)          #
+# All Rights Reserved.                                                         #
+#                                                                              #
+################################################################################
+
+include_directories(${CMAKE_CURRENT_SOURCE_DIR})
+
+set(SRC module.cpp)
+set(HDRS module.h)
+
+yarp_prepare_plugin(myModuleBar
+                    CATEGORY rfmodule
+                    TYPE moduleBar
+                    INCLUDE module.h)
+
+if(ENABLE_myModules_myModuleBar)
+  yarp_add_plugin(module_Bar ${SRC} ${HDRS})
+  target_link_libraries(module_Bar YARP::YARP_OS)
+endif()
+
+#create the executable target
+add_executable(myModuleBarExe main.cpp ${SRC} ${HDRS})
+target_link_libraries(myModuleBarExe YARP::YARP_OS YARP::YARP_init)
