+++ example/RFModulePlugin/testmoduleBar/myModuleBar.ini
@@ -1,0 +1,4 @@
+[plugin myModuleBar]
+type rfmodule
+name myModuleBar
+library module_Bar
