+++ example/RFModulePlugin/testmoduleFoo/myModuleFoo.ini
@@ -1,0 +1,4 @@
+[plugin myModuleFoo]
+type rfmodule
+name myModuleFoo
+library module_Foo
